pre-format transaction output
I want to separate the date when the money was used and when it was recorded. And post to ynab when it was used and not recorded, to get real time view of when money was moved.
I've also formatted the payee to match with payees in ynab and then allow it to match with category specified related to the payee in ynab. 
But this is just the transaction output so this formatting needs to be applied also to the creation of transaction.